requirements.txt
# FastAPI and Web Framework
fastapi[all]~=0.115.12
fastapi-mcp~=0.4.0

# Database
SQLAlchemy~=2.0.41
alembic~=1.16.4
psycopg2-binary~=2.9.10
psycopg~=3.2.10
psycopg-pool~=3.2.6

# Document Processing
Jinja2~=3.1.6
docxtpl~=0.19.1
python-docx~=1.1.2
python-docx-replace~=0.4.4

# LangChain and LangGraph
langchain~=0.3.27
langchain-core~=0.3.76
langchain-community~=0.3.29
langchain-google-community~=2.0.7
langchain-google-genai~=2.1.9
langchain-mcp-adapters~=0.1.10
langchain-tavily~=0.2.11
langgraph~=0.6.5
langgraph-checkpoint~=2.1.1
langgraph-checkpoint-postgres~=2.0.24
langgraph-prebuilt~=0.6.4
langgraph-supervisor~=0.0.29
langmem~=0.0.29

# Google Cloud Services
google-cloud-documentai~=3.5.0
google-adk~=1.3.0
google-genai~=1.20.0
google-api-python-client~=2.172.0
google-auth-oauthlib~=1.2.2

# AWS Services
boto3~=1.38.23
botocore~=1.38.23

# Authentication and Security
PyJWT~=2.10.1
passlib~=1.7.4
bcrypt~=3.2.2
cryptography~=45.0.4

# Utilities
python-dotenv~=1.1.0
phonenumbers~=9.0.5
pydantic~=2.11.4
protobuf~=6.31.1
email-validator~=2.2.0